pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "binterint"
version = "0.1.0"
description = "Headless TUI Interaction & Screenshot Binary"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Maheshwar Anup", email = "maheshwaranup@gmail.com"}
]
dependencies = [
    "ptyprocess>=0.7.0; sys_platform != 'win32'",
    "pywinpty>=2.0.0; sys_platform == 'win32'",
    "pyte>=0.8.0",
    "Pillow>=10.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
llm = [
    "google-generativeai",
    "openai",
]
dev = [
    "pytest",
    "black",
    "isort",
]

[project.scripts]
binterint = "binterint.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["binterint*"]

[tool.setuptools.package-data]
binterint = ["fonts/*.ttf"]
